// server/src/enums/statusCode.ts
export enum StatusCode {
    OK = 200,
    CREATED = 201,
    NO_CONTENT = 204,
  
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    CONFLICT = 409,
    TOO_MANY_REQUESTS = 429,
  
    INTERNAL_SERVER_ERROR = 500,
    NOT_IMPLEMENTED = 501,
    SERVICE_UNAVAILABLE = 503,
    METHOD_NOT_ALLOWED=504,
  }
  